backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
google-auth==2.23.0
requests==2.31.0
gunicorn==21.2.0